separate file in ruby method
