+++ ruby/api/book/batch_book_api.rb
@@ -1,0 +1,28 @@
+# frozen_string_literal: true
+
+require "#{File.dirname(__FILE__)}/../book_api"
+
+# Service to get books in batch
+class BatchBookApi < BookApi
+  # batch get item
+  # max 100 items
+  # retrieve items in paralel
+  # no order
+
+  def books
+    response = benmark('Batch get item in:') do
+      @client.batch_get_item({
+                               request_items: {
+                                 TABLE_NAME => {
+                                   keys: [
+                                     {
+                                       'isbn' => '1933988673'
+                                     }
+                                   ]
+                                 }
+                               }
+                             })
+    end
+    response.responses
+  end
+end
